Time: 191 ms (5.02%), Space: 123.8 MB (13.47%) - LeetHub

+++ 2615-sum-of-distances/2615-sum-of-distances.cpp
@@ -4,7 +4,7 @@ public:
         // arr[i] = (frq[nums[i]] - 1) * i - sum + i 
         int n = nums.size();
         vector <long long> arr(n, 0);
-        map <int, int> frq;
+        map <int, long long> frq;
         map <int, long long> sum;
         for(int i = 0; i < n; i ++){
             arr[i] += frq[nums[i]] * i - sum[nums[i]];
